exercise2_service/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
django-environ==0.11.2
drf-spectacular==0.27.0
setuptools==69.0.0
